Weather_bot/requirements.txt
aiogram==3.20.0
python-decouple==3.8
redis==5.0.0 
httpx==0.24.1
asyncpg==0.30.0
requests~=2.32.3
pillow~=11.2.1